Added Tabs Sample
