+++ Assets/Sandbox/Tabs/Editor/RuntimeUiInspector.cs
@@ -4,7 +4,7 @@ using UnityEngine.UIElements;
 
 namespace QuickEye.UIToolkit.Sandbox
 {
-    [CustomEditor(typeof(RuntimeUi))]
+    //[CustomEditor(typeof(RuntimeUi))]
     public class RuntimeUiInspector : UnityEditor.Editor
     {
         [SerializeField]
@@ -14,7 +14,6 @@ namespace QuickEye.UIToolkit.Sandbox
         private VisualTreeAsset inspectorTemplate;
 
         private RuntimeUiInspectorView _inspectorView;
-
 
         public override VisualElement CreateInspectorGUI()
         {
