+++ Packages/com.quickeye.ui-toolkit-plus/Samples/Tabs.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0c644d5c0b78d41beaa8b43d15755f74
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
