+++ Packages/com.quickeye.ui-toolkit-plus/Samples/Tabs/Editor/QuickEye.UIToolkit.Samples.Tabs.Editor.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 443fa65a253b44846aa0980c29ebc7d5
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
