+++ Packages/com.quickeye.ui-toolkit-plus/Samples/Tabs/Runtime/QuickEye.UIToolkit.Samples.Tabs.asmdef
@@ -1,0 +1,16 @@
+{
+    "name": "QuickEye.UIToolkit.Samples.Tabs",
+    "rootNamespace": "QuickEye.UIToolkit.Samples.Tabs",
+    "references": [
+        "GUID:9861c8596b4de1544a09a56d39ef2572"
+    ],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
