+++ Packages/com.quickeye.ui-toolkit-plus/Samples/Tabs/Runtime/QuickEye.UIToolkit.Samples.Tabs.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: db7f11f4da11a4c44a04260a83b3b085
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
