+++ Packages/com.quickeye.ui-toolkit-plus/Samples/Tabs/Runtime/TabView.cs
@@ -1,18 +1,18 @@
 using UnityEngine;
 using UnityEngine.UIElements;
 
-namespace QuickEye.UIToolkit.Sandbox
+namespace QuickEye.UIToolkit.Samples.Tabs
 {
     public partial class TabView
     {
         public TabView(VisualElement root)
         {
             AssignQueryResults(root);
-            horTabDropdown1.BeforeMenuShow += PopulateDropdown;
-            vertTabDropdown1.BeforeMenuShow += PopulateDropdown;
+            horTabDropdown1.BeforeMenuShow += PopulateSceneDropdown;
+            vertTabDropdown1.BeforeMenuShow += PopulateSceneDropdown;
         }
 
-        private static void PopulateDropdown(IGenericMenu menu)
+        private static void PopulateSceneDropdown(IGenericMenu menu)
         {
             menu.AddItem("Hello, World!", false, () => { Debug.Log($"Hello, World!"); });
         }
