+++ Packages/com.quickeye.ui-toolkit-plus/Samples/Tabs/UI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a64924cc723704177b491b226a718a4d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
