+++ Packages/com.quickeye.ui-toolkit-plus/Samples/Tabs/UI/Tabs.asmref
@@ -1,0 +1,3 @@
+{
+    "reference": "GUID:db7f11f4da11a4c44a04260a83b3b085"
+}
