+++ Packages/com.quickeye.ui-toolkit-plus/Samples/Tabs/UI/Tabs.asmref.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a795603c0676d4291870ddf8e76730c1
+AssemblyDefinitionReferenceImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
